Update Current Projects and Things I Read on homepage
Prior to this commit, the projects listed on the home page were badly
out of date and did not reflect the current set of projects I am
working on. Also, the Things I Read section contained blogs that are
not longer maintained.

This commit updates the project list and reading links on the homepage.