Fix anchor tags in new home page
Prior to this commit, none of the anchor tags in the home screen were
able to resolve to an address because they were left empty. This is
because I had to build the thing while I was on an airplane with no
internet. Fun.

This commit populates the anchor tags with real URLs.